# openapi.yaml
...
openapi: "3.0.0"
info:
  version: v1
  title: Bedrock API
  description: "Bedrock api k8s"
paths:
...
  /clients/{clientId}/tools/toolbelt:
    get:
      summary: Get info toolbelt box
      description: the clientId should be created previously
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: string
      tags:
        - Toolbelts
      responses:
        '200':
          description: toolbelt status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Toolbelt'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      summary: Create toolbelt box
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: string
      tags:
        - Toolbelts
      responses:
        '201':
          description: toolblet created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Toolbelt'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Delete toolbelt box
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: string
      tags:
        - Toolbelts
      responses:
        '200':
          description: toolbelt deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Toolbelt'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
...
    Toolbelt:
      properties:
        clientId:
          type: string
        url:
          type: string
        message:
          type: string
    Error:
      required:
        - code
        - msg
      properties:
        code:
          type: integer
          format: int32
        msg:
          type: string